fix(analytics): rename engine to analytics.service.ts, harden env resolution, gate PII from packaged logs
Addresses a batch of PR review findings on the analytics abstraction layer:

- Rename analytics.service-host.ts -> analytics.service.ts. The *.service-host.ts
  convention is reserved for modules paired with a shared *.service.ts/*.service-model.ts;
  this engine has no such pair, so the plain *.service.ts name (matching
  local-oauth.service.ts's precedent) is correct.
- Redesign getSelectedServer() to wait for the S/R server-target data provider via
  networkObjectStatusService.waitForNetworkObject() (event-driven, resolves as soon as
  the provider registers) instead of racing a fixed timeout, and split its failure
  handling into three independently-logged cases (registration timeout, provider
  vanished after registering, read failure) so a stalled resolution can be diagnosed
  from the log instead of guessed at.
- Split ConsoleAnalyticsProvider's log output into a safe `info`-level summary
  (name/timestamp/environment only) and a `debug`-level full event including
  `properties`, since `properties` is arbitrary caller-supplied data that tends to
  accumulate user/project identifiers, and `info` is what a packaged build's
  persistent log file captures by default.
- Rename PARATEXT_ANALYTICS_TEST_OVERRIDE -> PT_ANALYTICS_TEST_OVERRIDE for consistency
  with PT_STARTUP_MARKS, with strict `=== 'true'` matching; document it in README.md.
- Fix drainQueue()'s falsy-item bug (`while (next)` -> `while (next !== undefined)`) and
  correct its TSDoc ("per-iteration copy" -> "per-iteration binding").
- Add TSDoc to AnalyticsProvider.send(), initialize(), and trackEvent(); correct a stale
  extension-host.ts comment left over from the waitForDuration -> waitForNetworkObject
  change.
- Record the extension-owned-data-provider dependency (core's analytics.service.ts
  reading paratextRegistration's data provider) as an explicit ADR-0013 addendum, since
  it only typechecks via an incidental typeRoots property and shouldn't be read as a
  general precedent for core depending on extension-provided data providers.
