+++ src/extension-host/extension-host.ts
@@ -21,7 +21,7 @@ import { gracefulShutdownMessage } from '@node/models/interprocess-messages.mode
 import { killChildProcessesFromExtensions } from '@extension-host/services/create-process.service';
 import { initialize as initializeDatabaseService } from '@extension-host/services/database.service-host';
 import { startLocalOAuthServer } from '@extension-host/services/local-oauth.service';
-import * as analyticsService from '@extension-host/services/analytics.service-host';
+import * as analyticsService from '@extension-host/services/analytics.service';
 import { markStartup } from '@shared/utils/startup-timing.util';
 import { STARTUP_MARK_PROCESS_START } from '@shared/data/platform.data';
 
@@ -94,6 +94,13 @@ process.on('unhandledRejection', (reason) => {
     // Analytics initialization must not gate extension-host startup — it can take up to
     // ENVIRONMENT_RESOLUTION_TIMEOUT_MS if the Send/Receive server-target lookup is slow.
     // trackEvent()'s `unresolved` queue tolerates this finishing at any point.
+    // This .catch() is unreachable today: resolveEnvironment() never rejects -- its async step,
+    // getSelectedServer(), catches every failure internally (timeout, vanished provider, read
+    // failure) and resolves to `undefined` rather than throwing -- and the rest of initialize()'s
+    // body is synchronous and already self-guarded. It's a defensive guard, not live error
+    // handling -- kept so that if a future change to analytics.service.ts ever breaks that
+    // never-rejects guarantee, it fails safe (logged) instead of raising an unhandled promise
+    // rejection here.
     analyticsService.initialize().catch((error) => {
       logger.error(`Analytics: failed to initialize: ${String(error)}`);
     });
