+++ src/extension-host/services/analytics-providers/console-analytics.provider.test.ts
@@ -18,7 +18,7 @@ beforeEach(() => {
   mocks.warn.mockClear();
 });
 
-test('a test-environment provider logs the event with a Test prefix, including its real properties and timestamp', async () => {
+test('a test-environment provider logs a safe Test-prefixed summary at info, and the full event (properties included) at debug', async () => {
   const { ConsoleAnalyticsProvider } = await import(
     '@extension-host/services/analytics-providers/console-analytics.provider'
   );
@@ -32,14 +32,24 @@ test('a test-environment provider logs the event with a Test prefix, including i
   });
 
   expect(mocks.info).toHaveBeenCalledTimes(1);
-  const [message] = mocks.info.mock.calls[0];
-  expect(message).toContain('Test: ');
-  expect(message).toContain('"name":"app_launch"');
-  expect(message).toContain('"version":"1.2.3"');
-  expect(message).toContain('"timestamp":1700000000000');
+  const [infoMessage] = mocks.info.mock.calls[0];
+  expect(infoMessage).toContain('Test: ');
+  expect(infoMessage).toContain('"name":"app_launch"');
+  expect(infoMessage).toContain('"timestamp":1700000000000');
+  // The info line -- visible in packaged builds -- must never carry arbitrary caller-supplied
+  // properties, since those tend to accumulate user/project identifiers over time and would land
+  // verbatim in a real user's persistent log file.
+  expect(infoMessage).not.toContain('version');
+  expect(infoMessage).not.toContain('1.2.3');
+
+  expect(mocks.debug).toHaveBeenCalledTimes(1);
+  const [debugMessage] = mocks.debug.mock.calls[0];
+  expect(debugMessage).toContain('Test');
+  expect(debugMessage).toContain('"name":"app_launch"');
+  expect(debugMessage).toContain('"version":"1.2.3"');
 });
 
-test('a production-environment provider logs the event with a Production prefix, including its real name and timestamp', async () => {
+test('a production-environment provider logs a safe Production-prefixed summary at info, and the full event at debug', async () => {
   const { ConsoleAnalyticsProvider } = await import(
     '@extension-host/services/analytics-providers/console-analytics.provider'
   );
@@ -52,13 +62,18 @@ test('a production-environment provider logs the event with a Production prefix,
   });
 
   expect(mocks.info).toHaveBeenCalledTimes(1);
-  const [message] = mocks.info.mock.calls[0];
-  expect(message).toContain('Production: ');
-  expect(message).toContain('"name":"subscription_renewed"');
-  expect(message).toContain('"timestamp":1705000000000');
+  const [infoMessage] = mocks.info.mock.calls[0];
+  expect(infoMessage).toContain('Production: ');
+  expect(infoMessage).toContain('"name":"subscription_renewed"');
+  expect(infoMessage).toContain('"timestamp":1705000000000');
+
+  expect(mocks.debug).toHaveBeenCalledTimes(1);
+  const [debugMessage] = mocks.debug.mock.calls[0];
+  expect(debugMessage).toContain('Production');
+  expect(debugMessage).toContain('"name":"subscription_renewed"');
 });
 
-test('a test provider given a production-tagged event still logs its real data, but warns about the mismatch', async () => {
+test('a test provider given a production-tagged event still logs it at both levels, but warns about the mismatch', async () => {
   const { ConsoleAnalyticsProvider } = await import(
     '@extension-host/services/analytics-providers/console-analytics.provider'
   );
@@ -72,15 +87,20 @@ test('a test provider given a production-tagged event still logs its real data, 
   });
 
   expect(mocks.warn).toHaveBeenCalledWith(expect.stringContaining("tagged as 'production'"));
+
   expect(mocks.info).toHaveBeenCalledTimes(1);
-  const [message] = mocks.info.mock.calls[0];
-  expect(message).toContain('Test: ');
-  expect(message).toContain('"name":"project_synced"');
-  expect(message).toContain('"source":"test-harness"');
-  expect(message).toContain('"timestamp":1710000000000');
+  const [infoMessage] = mocks.info.mock.calls[0];
+  expect(infoMessage).toContain('Test: ');
+  expect(infoMessage).toContain('"name":"project_synced"');
+  expect(infoMessage).not.toContain('source');
+  expect(infoMessage).not.toContain('test-harness');
+
+  expect(mocks.debug).toHaveBeenCalledTimes(1);
+  const [debugMessage] = mocks.debug.mock.calls[0];
+  expect(debugMessage).toContain('"source":"test-harness"');
 });
 
-test('a production provider given a test-tagged event warns and refuses to log/transmit it', async () => {
+test('a production provider given a test-tagged event warns and refuses to log/transmit it at either level', async () => {
   const { ConsoleAnalyticsProvider } = await import(
     '@extension-host/services/analytics-providers/console-analytics.provider'
   );
@@ -94,4 +114,5 @@ test('a production provider given a test-tagged event warns and refuses to log/t
 
   expect(mocks.warn).toHaveBeenCalledWith(expect.stringContaining("tagged as 'test'"));
   expect(mocks.info).not.toHaveBeenCalled();
+  expect(mocks.debug).not.toHaveBeenCalled();
 });
