+++ src/extension-host/services/analytics.service.test.ts
@@ -1,0 +1,362 @@
+import { afterEach, beforeEach, expect, test, vi } from 'vitest';
+
+const mocks = vi.hoisted(() => ({
+  get: vi.fn(),
+  waitForNetworkObject: vi.fn(),
+  debug: vi.fn(),
+  info: vi.fn(),
+  warn: vi.fn(),
+  error: vi.fn(),
+}));
+
+vi.mock('@shared/services/data-provider.service', () => ({
+  dataProviderService: { get: mocks.get },
+}));
+vi.mock('@shared/services/network-object-status.service', () => ({
+  networkObjectStatusService: { waitForNetworkObject: mocks.waitForNetworkObject },
+}));
+vi.mock('@shared/services/logger.service', () => ({
+  __esModule: true,
+  default: { debug: mocks.debug, info: mocks.info, warn: mocks.warn, error: mocks.error },
+  logger: { debug: mocks.debug, info: mocks.info, warn: mocks.warn, error: mocks.error },
+}));
+
+/**
+ * Find the console-provider's safe `info`-level summary line for a given environment label, if one
+ * was logged. Deliberately excludes `properties` -- see `findFullEventLog` for that.
+ */
+function findSentLog(label: 'Test' | 'Production'): string | undefined {
+  return mocks.info.mock.calls
+    .map(([message]) => message)
+    .find((message) => message.startsWith(`${label}: `));
+}
+
+/**
+ * Find the console-provider's `debug`-level full-event line (properties included) for a given
+ * environment label, if one was logged.
+ */
+function findFullEventLog(label: 'Test' | 'Production'): string | undefined {
+  return mocks.debug.mock.calls
+    .map(([message]) => message)
+    .find((message) => message.startsWith(`${label} (full event): `));
+}
+
+beforeEach(() => {
+  vi.clearAllMocks();
+  vi.resetModules();
+  vi.unstubAllEnvs();
+  globalThis.isPackaged = true;
+  vi.stubEnv('NODE_ENV', 'production');
+  // Default: the data provider is already available. Tests exercising the wait/timeout/failure
+  // mechanics override this explicitly.
+  mocks.waitForNetworkObject.mockResolvedValue({
+    id: 'paratextRegistration.internetSettingsDataProvider',
+  });
+});
+
+afterEach(() => {
+  vi.useRealTimers();
+  // `vi.doMock` (used by the "failing provider send" test below) isn't undone by
+  // `vi.resetModules()` — it stays registered for the module path and would otherwise leak into
+  // every later test in this file that re-imports the module.
+  vi.doUnmock('@extension-host/services/analytics-providers/console-analytics.provider');
+});
+
+test('an event fired before initialize resolves is queued, then flushed as test once the override env var forces test', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'true');
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+
+  trackEvent('app_launch');
+  expect(mocks.debug).toHaveBeenCalledWith('Analytics event tracked: app_launch');
+  expect(findSentLog('Test')).toBeUndefined();
+  expect(mocks.get).not.toHaveBeenCalled();
+
+  await initialize();
+
+  const sentLog = findSentLog('Test');
+  expect(sentLog).toBeDefined();
+  expect(sentLog).toContain('"name":"app_launch"');
+});
+
+test('the override env var only activates on an exact "true" value, not any truthy-looking string', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'false');
+  mocks.get.mockResolvedValue({
+    getInternetSettings: vi.fn().mockResolvedValue({ selectedServer: 'Production' }),
+  });
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  await initialize();
+
+  // A non-'true' string must not short-circuit to test -- it should fall through to the real
+  // build-type/S-R-target checks, which in this packaged+Production setup resolve 'production'.
+  expect(findSentLog('Production')).toBeDefined();
+  expect(findSentLog('Test')).toBeUndefined();
+});
+
+test('an unpackaged dev build resolves to test without checking the S/R server target', async () => {
+  globalThis.isPackaged = false;
+  vi.stubEnv('NODE_ENV', 'development');
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  await initialize();
+
+  expect(mocks.get).not.toHaveBeenCalled();
+  expect(findSentLog('Test')).toBeDefined();
+});
+
+test('a packaged production build targeting a non-Production S/R server resolves to test', async () => {
+  mocks.get.mockResolvedValue({
+    getInternetSettings: vi.fn().mockResolvedValue({ selectedServer: 'QualityAssurance' }),
+  });
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  await initialize();
+
+  expect(mocks.get).toHaveBeenCalledWith('paratextRegistration.internetSettingsDataProvider');
+  expect(findSentLog('Test')).toBeDefined();
+});
+
+test('a packaged production build targeting the Production S/R server resolves to production', async () => {
+  mocks.get.mockResolvedValue({
+    getInternetSettings: vi.fn().mockResolvedValue({ selectedServer: 'Production' }),
+  });
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  await initialize();
+
+  expect(findSentLog('Production')).toBeDefined();
+  expect(findSentLog('Test')).toBeUndefined();
+});
+
+test('an S/R server lookup where the data provider never registers falls back to test after the timeout, with a warning logged', async () => {
+  vi.useFakeTimers();
+  // Simulate waitForNetworkObject's own real timeout behavior (AsyncVariable rejects once
+  // timeoutMs elapses) since networkObjectStatusService is mocked at the module boundary.
+  mocks.waitForNetworkObject.mockImplementation(
+    (_details: unknown, timeoutMs: number) =>
+      new Promise((_resolve, reject) => {
+        setTimeout(() => reject(new Error('Timeout reached when waiting to settle')), timeoutMs);
+      }),
+  );
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  const initializePromise = initialize();
+
+  await vi.advanceTimersByTimeAsync(30_000);
+  await initializePromise;
+
+  // The debug log names this specifically as a timeout, distinct from the "provider vanished" and
+  // "failed to read settings" cases below -- so a real diagnosis doesn't have to guess which of
+  // the three happened.
+  expect(mocks.debug).toHaveBeenCalledWith(expect.stringContaining('timed out waiting'));
+  expect(mocks.warn).toHaveBeenCalledWith(expect.stringContaining('test environment'));
+  expect(findSentLog('Test')).toBeDefined();
+});
+
+test('the data provider vanishing after a successful wait logs a distinct debug message and falls back to test', async () => {
+  // waitForNetworkObject resolves (the provider was seen registering), but a subsequent get()
+  // finds nothing -- a real, distinct race, not a timeout and not a read failure.
+  mocks.get.mockResolvedValue(undefined);
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  await initialize();
+
+  expect(mocks.debug).toHaveBeenCalledWith(expect.stringContaining('vanished'));
+  expect(findSentLog('Test')).toBeDefined();
+});
+
+test('an S/R server lookup where the data provider registers partway through the wait resolves production correctly, without waiting the full budget', async () => {
+  vi.useFakeTimers();
+
+  // Simulate the data provider registering partway through the wait window -- the event-driven
+  // wait should catch this as soon as it happens, well before the full 30s budget.
+  const delayMs = 5000;
+  mocks.waitForNetworkObject.mockImplementation(
+    () =>
+      new Promise((resolve) => {
+        setTimeout(
+          () => resolve({ id: 'paratextRegistration.internetSettingsDataProvider' }),
+          delayMs,
+        );
+      }),
+  );
+  mocks.get.mockResolvedValue({
+    getInternetSettings: vi.fn().mockResolvedValue({ selectedServer: 'Production' }),
+  });
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  const initializePromise = initialize();
+
+  // Only advance to the registration delay, not the full 30s timeout -- proves this resolves as
+  // soon as the object appears rather than always paying the full wait budget.
+  await vi.advanceTimersByTimeAsync(delayMs);
+  await initializePromise;
+
+  expect(findSentLog('Production')).toBeDefined();
+  expect(findSentLog('Test')).toBeUndefined();
+});
+
+test('an error while waiting for the data provider resolves promptly instead of waiting out the full timeout', async () => {
+  vi.useFakeTimers();
+  mocks.waitForNetworkObject.mockRejectedValue(
+    new Error('network object status service unavailable'),
+  );
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  const initializePromise = initialize();
+
+  // Only a small amount of fake time is needed for the rejection to be caught and resolved to
+  // 'test' -- an immediate rejection is caught immediately, unlike the old waitForDuration-based
+  // design where a rejection would have waited out the whole timeout regardless.
+  await vi.advanceTimersByTimeAsync(50);
+  await initializePromise;
+
+  expect(mocks.debug).toHaveBeenCalledWith(expect.stringContaining('timed out waiting'));
+  expect(findSentLog('Test')).toBeDefined();
+});
+
+test('a failure reading settings from an already-found data provider logs a distinct debug message and falls back to test', async () => {
+  mocks.get.mockResolvedValue({
+    getInternetSettings: vi.fn().mockRejectedValue(new Error('getInternetSettings failed')),
+  });
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  trackEvent('app_launch');
+  await initialize();
+
+  expect(mocks.debug).toHaveBeenCalledWith(expect.stringContaining('failed to read'));
+  expect(findSentLog('Test')).toBeDefined();
+});
+
+test('an event fired after initialize has already resolved is stamped and flushed immediately', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'true');
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  await initialize();
+  mocks.debug.mockClear();
+
+  trackEvent('second_event', { count: 2 });
+
+  const sentLog = findSentLog('Test');
+  expect(sentLog).toBeDefined();
+  expect(sentLog).toContain('"name":"second_event"');
+
+  const fullEventLog = findFullEventLog('Test');
+  expect(fullEventLog).toBeDefined();
+  expect(fullEventLog).toContain('"name":"second_event"');
+  expect(fullEventLog).toContain('"count":2');
+});
+
+test('a failing provider send is caught and logged, without initialize or trackEvent throwing', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'true');
+  vi.doMock('@extension-host/services/analytics-providers/console-analytics.provider', () => ({
+    ConsoleAnalyticsProvider: class {
+      constructor(private readonly environment: string) {}
+
+      async send(): Promise<void> {
+        throw new Error(`boom (${this.environment})`);
+      }
+    },
+  }));
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+
+  trackEvent('app_launch');
+  await expect(initialize()).resolves.toBeUndefined();
+
+  // Let the fire-and-forget `.catch()` microtask run before asserting on it
+  await new Promise((resolve) => {
+    setTimeout(resolve, 0);
+  });
+
+  expect(mocks.error).toHaveBeenCalledWith(expect.stringContaining('boom'));
+});
+
+test('trackEvent drops a non-serializable property and warns immediately, but still sends the rest of the event', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'true');
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  await initialize();
+  mocks.warn.mockClear();
+
+  const circular: Record<string, unknown> = {};
+  circular.self = circular;
+
+  trackEvent('bad_properties_event', { valid: 'ok', circularRef: circular });
+
+  // Warned at trackEvent() call time, before the event is ever queued or sent
+  expect(mocks.warn).toHaveBeenCalledWith(
+    expect.stringContaining("dropping non-serializable property 'circularRef'"),
+  );
+
+  const sentLog = findSentLog('Test');
+  expect(sentLog).toBeDefined();
+  expect(sentLog).toContain('"name":"bad_properties_event"');
+
+  const fullEventLog = findFullEventLog('Test');
+  expect(fullEventLog).toBeDefined();
+  expect(fullEventLog).toContain('"valid":"ok"');
+  expect(fullEventLog).not.toContain('circularRef');
+});
+
+test('trackEvent drops properties whose values are functions, symbols, or undefined -- JSON.stringify does not throw for these, it just silently omits them', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'true');
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  await initialize();
+  mocks.warn.mockClear();
+
+  trackEvent('bad_properties_event', {
+    valid: 'ok',
+    someFunction: () => {},
+    someSymbol: Symbol('x'),
+    explicitlyUndefined: undefined,
+  });
+
+  expect(mocks.warn).toHaveBeenCalledWith(
+    expect.stringContaining("dropping non-serializable property 'someFunction'"),
+  );
+  expect(mocks.warn).toHaveBeenCalledWith(
+    expect.stringContaining("dropping non-serializable property 'someSymbol'"),
+  );
+  expect(mocks.warn).toHaveBeenCalledWith(
+    expect.stringContaining("dropping non-serializable property 'explicitlyUndefined'"),
+  );
+
+  const fullEventLog = findFullEventLog('Test');
+  expect(fullEventLog).toBeDefined();
+  expect(fullEventLog).toContain('"valid":"ok"');
+  expect(fullEventLog).not.toContain('someFunction');
+  expect(fullEventLog).not.toContain('someSymbol');
+  expect(fullEventLog).not.toContain('explicitlyUndefined');
+});
+
+test('trackEvent omits properties entirely when none of them survive sanitization', async () => {
+  vi.stubEnv('PT_ANALYTICS_TEST_OVERRIDE', 'true');
+
+  const { initialize, trackEvent } = await import('@extension-host/services/analytics.service');
+  await initialize();
+
+  const circular: Record<string, unknown> = {};
+  circular.self = circular;
+
+  trackEvent('all_properties_bad', { onlyBad: circular });
+
+  const sentLog = findSentLog('Test');
+  expect(sentLog).toBeDefined();
+  expect(sentLog).toContain('"name":"all_properties_bad"');
+
+  const fullEventLog = findFullEventLog('Test');
+  expect(fullEventLog).toBeDefined();
+  expect(fullEventLog).toContain('"name":"all_properties_bad"');
+  expect(fullEventLog).not.toContain('"properties"');
+});
